Created a helper class to simplify code
