+++ WinUIGallery/SettingsPage.xaml.cs
@@ -20,6 +20,7 @@ using Microsoft.UI.Xaml.Input;
 using Microsoft.UI.Xaml.Navigation;
 using WinRT;
 using System.Runtime.InteropServices;
+using WinUIGallery.DesktopWap.Helper;
 
 #if UNIVERSAL
 using Windows.UI.ViewManagement;
@@ -90,12 +91,6 @@ namespace AppUIBasics
             }
         }
 
-        [ComImport, Guid("EECDBF0E-BAE9-4CB6-A68E-9598E1CB57BB"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-        private interface IWindowNative
-        {
-            IntPtr WindowHandle { get; }
-        };
-
         private void OnThemeRadioButtonChecked(object sender, RoutedEventArgs e)
         {
             var selectedTheme = ((RadioButton)sender)?.Tag?.ToString();
@@ -130,20 +125,7 @@ namespace AppUIBasics
                 }
             }
 
-            // to trigger repaint tracking task id 38044406
-            var native = App.StartupWindow.As<IWindowNative>();
-            var hwnd = native.WindowHandle;
-            var activeWindow = Win32.GetActiveWindow();
-            if (hwnd == activeWindow)
-            {
-                Win32.SendMessage(hwnd, Win32.WM_ACTIVATE, Win32.WA_INACTIVE, IntPtr.Zero);
-                Win32.SendMessage(hwnd, Win32.WM_ACTIVATE, Win32.WA_ACTIVE, IntPtr.Zero);
-            }
-            else
-            {
-                Win32.SendMessage(hwnd, Win32.WM_ACTIVATE, Win32.WA_ACTIVE, IntPtr.Zero);
-                Win32.SendMessage(hwnd, Win32.WM_ACTIVATE, Win32.WA_INACTIVE, IntPtr.Zero);
-            }
+            TitlebarHelper.triggerTitleBarRepaint();
 
         }
